+++ app/codecov.yml
@@ -1,0 +1,2 @@
+fixes:
+  - "server/::app/server/"
