--- app/test_app/helpers/misc_tasker.py
@@ -1,5 +1,0 @@
-class MockMiscTasker(object):
-
-    @staticmethod
-    def set_ip_location(*args, **kwargs):
-        pass
