+++ app/test_app/unit/utils/ussd/kenya/test_exchange_state.py
@@ -5,8 +5,8 @@ from faker import Faker
 import json
 from server import db
 
-from helpers.factories import UserFactory, UssdSessionFactory, OrganisationFactory
-from helpers.ussd_utils import make_kenyan_phone
+from helpers.model_factories import UserFactory, UssdSessionFactory, OrganisationFactory
+from helpers.utils import make_kenyan_phone
 from server.utils.ussd.kenya_ussd_state_machine import KenyaUssdStateMachine
 from server.models.user import User
