+++ app/test_app/unit/utils/ussd/kenya/test_send_state.py
@@ -4,8 +4,8 @@ from faker.providers import phone_number
 from faker import Faker
 import json
 
-from helpers.factories import UserFactory, UssdSessionFactory, TokenFactory, OrganisationFactory
-from helpers.ussd_utils import make_kenyan_phone, fake_transfer_mapping
+from helpers.model_factories import UserFactory, UssdSessionFactory, TokenFactory, OrganisationFactory
+from helpers.utils import make_kenyan_phone, fake_transfer_mapping
 from server import db
 from server.utils.ussd.kenya_ussd_state_machine import KenyaUssdStateMachine
 from server.models.user import User
