--- eth_worker/_beat_starter.sh
@@ -1,7 +1,0 @@
-#!/usr/bin/env bash
-
-#echo cheese
-
-celery -A worker beat --loglevel=WARNING
-
-#-A worker worker --loglevel=INFO --concurrency=500 --pool=eventlet
