+++ eth_worker/codecov.yml
@@ -1,3 +1,2 @@
 fixes:
-  - "server/::app/server/"
   - "::eth_worker/"
